// lib/api/model/getTransactionsParams.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Jeffries Homeapp API
 * Backend REST API for Jeffries Homeapp.
 * OpenAPI spec version: 1.0
 */

export type GetTransactionsParams = {
/**
 * User ID
 */
userId: string;
/**
 * Exclude internal transactions
 */
excludeIntern?: boolean;
/**
 * Only show reversed transactions
 */
onlyStorneringen?: boolean;
/**
 * Transaction code
 */
codeFilter?: string;
/**
 * IBAN filter
 */
ibanFilter?: string;
/**
 * Category filter
 */
categorieFilter?: string;
/**
 * Direction (in/uit)
 */
richting?: string;
/**
 * Start date (YYYY-MM-DD)
 */
datumVan?: string;
/**
 * End date (YYYY-MM-DD)
 */
datumTot?: string;
/**
 * Search query
 */
zoekterm?: string;
/**
 * Minimum amount
 */
minBedrag?: number;
/**
 * Maximum amount
 */
maxBedrag?: number;
/**
 * Year filter
 */
jaarFilter?: string;
/**
 * Limit count
 */
limit?: number;
/**
 * Offset count
 */
offset?: number;
};