// lib/api/model/getTransactionsStatsParams.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Jeffries Homeapp API
 * Backend REST API for Jeffries Homeapp.
 * OpenAPI spec version: 1.0
 */

export type GetTransactionsStatsParams = {
/**
 * User ID
 */
userId: string;
/**
 * IBAN filter
 */
ibanFilter?: string;
/**
 * Year filter
 */
jaarFilter?: string;
};