// lib/api/model/modelTransaction.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Jeffries Homeapp API
 * Backend REST API for Jeffries Homeapp.
 * OpenAPI spec version: 1.0
 */

export interface ModelTransaction {
  bedrag?: number;
  categorie?: string;
  code?: string;
  datum?: string;
  id?: string;
  is_interne_overboeking?: boolean;
  omschrijving?: string;
  oorsp_bedrag?: number;
  oorsp_munt?: string;
  reden_retour?: string;
  referentie?: string;
  rekening_iban?: string;
  saldo_na_trn?: number;
  tegenpartij_naam?: string;
  tegenrekening_iban?: string;
  user_id?: string;
  volgnr?: string;
}