// lib/api/model/storeTransactionStats.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Jeffries Homeapp API
 * Backend REST API for Jeffries Homeapp.
 * OpenAPI spec version: 1.0
 */
import type { StoreCategorieBreakdown } from './storeCategorieBreakdown';
import type { StoreMaandInUit } from './storeMaandInUit';
import type { StoreMaandSaldo } from './storeMaandSaldo';
import type { StoreTopMerchant } from './storeTopMerchant';
import type { StoreTransactionStatsHuidigSaldoPerIban } from './storeTransactionStatsHuidigSaldoPerIban';
import type { StoreTransactionStatsSaldoPeildatumPerIban } from './storeTransactionStatsSaldoPeildatumPerIban';

export interface StoreTransactionStats {
  aantalAlleTxs: number;
  aantalCategorieen: number;
  aantalTxs: number;
  gemiddeldIn: number;
  gemiddeldUit: number;
  /** Echte bankbalans */
  huidigSaldo: number;
  huidigSaldoPerIban: StoreTransactionStatsHuidigSaldoPerIban;
  ibannen: string[];
  inPerCategorie: StoreCategorieBreakdown[];
  inUitPerMaand: StoreMaandInUit[];
  jaren: string[];
  laatsteSaldoPeildatum?: string;
  maanden: string[];
  nettoStroom: number;
  saldoPeildatumPerIban: StoreTransactionStatsSaldoPeildatumPerIban;
  /** Grafieken */
  saldoPerMaand: StoreMaandSaldo[];
  /** Overige */
  storneringen: number;
  topMerchants: StoreTopMerchant[];
  /** Kasstromen */
  totaalIn: number;
  totaalUit: number;
  /** Categorieën */
  uitPerCategorie: StoreCategorieBreakdown[];
}